Settings now throw RuntimeException if the input handler is missing

+++ src/cz/autoclient/settings/Settings.java
@@ -33,22 +33,26 @@ import javax.swing.JTextField;
  *
  * @author Jakub
  */
-public class Settings {
+public class Settings implements java.io.Serializable {
+  /** 
+   * Prevent any incompatible class bullshit errors
+   */
+  private static final long serialVersionUID = 666;
   //Settings can be of any type
   protected Map<String, Object> settings = new HashMap<>();
   /**
    * Map of input fields that are mapped to certain settings. This allows to automatically update the settings both ways.
    */
-  protected final Map<String, Input> boundInputs = new HashMap<>();
+  protected transient final Map<String, Input> boundInputs = new HashMap<>();
   /**
    * Should never change from true to false! Indicates that settings have changed and should be saved in file.
    */
-  private boolean changed = false;
+  private transient boolean changed = false;
   /**
    * If using encrypted settings, you must set the encryptor/decryptor first! Otherwise, 
    * {@link java.lang.IllegalStateException} will be thrown.
    */
-  protected SecureSettings encryptor;
+  protected transient SecureSettings encryptor;
 
 
   /**
@@ -109,7 +113,19 @@ public class Settings {
   public Object getSetting(String name) {
     return settings.get(name);
   }
-  
+  /** Retrieves value only if it's instance of given type. 
+   * 
+   * @param <T> 
+   * @param name name of the setting
+   * @param type Class of required return type.
+   * @return Setting cast to required type or null if this type is not present.
+   */
+  public <T> T getSetting(String name, Class<T> type) {
+    Object val = settings.get(name);
+    if(val!=null && type.isInstance(val))
+      return (T)val;
+    return null;
+  }
   public Object getEncrypted(String name) throws InvalidPasswordException {
     Object s = settings.get(name);
     if(s instanceof EncryptedSetting) {
@@ -129,11 +145,11 @@ public class Settings {
       throw new IllegalStateException("The SecureSettings class must be initialised before using encrypted"
           + "settings!"); 
     }
-    System.out.println("ENCRYPTED: AutomatSettings[\""+name+"\"] = "+(value!=null?"[value hidden]":"null"));
+    System.out.println("ENCRYPTED: AutomatSettings[\""+name+"\"] = "+value/*(value!=null?"[value hidden]":"null")*/);
     changed = true;
     Object s = settings.get(name);
     if(s!=null && s instanceof EncryptedSetting) {
-      ((EncryptedSetting)s).encryptor = encryptor;
+      ((EncryptedSetting)s).setEncryptor(encryptor);
       ((EncryptedSetting)s).setDecryptedValue(value);
     }
     else {
@@ -145,10 +161,8 @@ public class Settings {
   
   public boolean empty(String name) {
     Object val = settings.get(name);
-    if(val!=null) {
-      return false;
-    }
-    return true;
+
+    return val==null;
   } 
   /** 
    * Check if setting with given name and type exists. This can be used to prevent 
@@ -172,21 +186,35 @@ public class Settings {
   public boolean exists(String name) {
     return settings.containsKey(name);
   }
+  /** 
+   * Return the size of the settings array.
+   * @return number of elements in the internal HashMap
+   */
+  public int size() {
+    return settings.size();
+  }
   /** Set setting to any object value. 
    * @param name setting name
    * @param value setting value
+   * @param force forces the setting to be written even if old value equals the new
    * @return old setting value. If the setting didn't exist before, returns null. Note that null is also valid setting value
    */
-  public Object setSetting(String name, Object value) {
+  public Object setSetting(String name, Object value, boolean force) {
     Object old = settings.get(name);
-    if(old==value || old!=null&&old.equals(value)) {
-      return old;   
+    if(!force) {
+      if(old==value || old!=null&&old.equals(value)) {
+        return old;   
+      }
     }
     changed = true;
     settings.put(name, value);
     System.out.println("AutomatSettings[\""+name+"\"] = "+(value!=null?value.toString():"null"));
     //Thread.dumpStack();
     return old;
+  }
+  
+  public Object setSetting(String name, Object value) {
+    return setSetting(name, value, false);
   }
   
   /** Set default value of that setting. This must be set every time after loading settings (it's not saved).
@@ -204,6 +232,15 @@ public class Settings {
     settings.put(name, value);
     System.out.println("default AutomatSettings[\""+name+"\"] = "+(value!=null?value.toString():"null"));
     return true;
+  }
+  /** Clear all settings.
+   * 
+   * @return number of values deleted from hashmap
+   */
+  public int clearSettings() {
+    int len = settings.size();
+    settings.clear();
+    return len;
   }
   /** Automatically update setting value as user types.
    * @param setting_name What is the name of associated setting?
@@ -240,10 +277,9 @@ public class Settings {
         verif
       );
     }
-    //Silent fail
     catch(NoSuchMethodException e) {
-      System.err.println("No handler for "+input.getClass().getName());
-      return;
+      //System.err.println("No handler for "+input.getClass().getName());
+      throw new NoInputHandlerException(e.getMessage(), input);
     }
     
     boundInputs.put(setting_name, in);
@@ -357,8 +393,6 @@ public class Settings {
   }
   public void loadSettingsFromBoundFields() {
     for (Map.Entry pair : boundInputs.entrySet()) {
-      
-      
       String name = (String)pair.getKey();
       Input input = (Input)pair.getValue();
       
@@ -371,43 +405,62 @@ public class Settings {
     }
   }
   
-  public void saveToFile(String path) throws IOException {
-    File f = new File(path);
-    f.getParentFile().mkdir();
+  public void saveToFile(File path) throws IOException {
+    
+    //File f = new File(path);
+    if(!path.getParentFile().exists()&&!path.getParentFile().mkdirs()) {
+      throw new IOException("Can't create folders in "+path.getAbsolutePath());
+    }
+    path.createNewFile();
+    //Ensure the settings will try to be saved encrypted
+    assignEncryptorToAllEncryptedSettings();
     
     OutputStream file = new FileOutputStream(path);
     OutputStream buffer = new BufferedOutputStream(file);
     ObjectOutput output = new ObjectOutputStream(buffer);
-    //System.out.println("Saving settings HashMap - "+settings.size()+" fields.");
+    System.out.println("Saving settings HashMap - "+settings.size()+" fields.");
     output.writeObject(settings);
     
     buffer.close();
     output.close();
   }
+  public void saveToFile(String path) throws IOException {
+    saveToFile(new File(path)); 
+  }
   public void saveToFile(String path, boolean forced) throws IOException {
+    saveToFile(new File(path), forced); 
+  }
+  public void saveToFile(File path, boolean forced) throws IOException {
     if(forced || changed)
       saveToFile(path);
   }
-  public void loadFromFile(String path) throws IOException {    
+  
+  public void loadFromFile(File path) throws IOException {    
     InputStream file = new FileInputStream(path);
     InputStream buffer = new BufferedInputStream(file);
     ObjectInput input = new ObjectInputStream (buffer);
     //Get the properties out of the object
     try {
       Map set = (Map)input.readObject();
-      //System.out.println(set.size()+" settings loaded from file.");
+      System.out.println(set.size()+" settings loaded from file.");
       this.settings.putAll(set);
     }
-    catch(ClassNotFoundException e) {};  
+    catch(ClassNotFoundException e) {
+    
+    };  
+  }
+  public void loadFromFile(String path) throws IOException {   
+    loadFromFile(new File(path)); 
   }
   
   /**Custom serialization - only write the hash map, everything else is irrelevant **/
-  /*private void writeObject(ObjectOutputStream oos) throws IOException {
+  private void writeObject(ObjectOutputStream oos) throws IOException {
+    assignEncryptorToAllEncryptedSettings();
     oos.writeObject(settings);
   }
   private void readObject(ObjectInputStream ois) throws ClassNotFoundException, IOException {
     settings = (HashMap<String, Object>)ois.readObject(); 
-  }*/
+  }
 
   public SecureSettings getEncryptor() {
     if(encryptor==null) {
@@ -415,8 +468,41 @@ public class Settings {
     }
     return encryptor;
   }
+  public boolean hasEncryptor() {
+    return encryptor!=null; 
+  }
 
   public void setEncryptor(SecureSettings encryptor) {
     this.encryptor = encryptor;
+  }
+  /** Loops through all secure settings and sets the own encryptor to them.
+   * 
+   */
+  private void assignEncryptorToAllEncryptedSettings() {
+    //Unlike other functions, this one is internal and fails silently
+    if(encryptor==null)
+      return;
+    for (Map.Entry pair : settings.entrySet()) {
+      if(pair.getValue() instanceof EncryptedSetting) {
+         encryptor.encrypt((EncryptedSetting)pair.getValue());
+      }
+    }
+  }
+  /** If this method is called, all settings will be decrypted. After that, the
+   *  any encryption keys can be changed and the encrypted data will not be lost.
+   * 
+   */
+  public void prepareForEncryptionChange() {
+    if(encryptor==null) {
+      throw new IllegalStateException("The SecureSettings class must be initialised before using encrypted"
+          + "settings!"); 
+    }
+    for (Map.Entry pair : settings.entrySet()) {
+      if(pair.getValue() instanceof EncryptedSetting) {
+         EncryptedSetting e = (EncryptedSetting)pair.getValue();
+         encryptor.decrypt(e);
+         e.forgetEncryptedValue();
+      }
+    }
   }
 }
